requirements.txt
pillow
opencv-python
jinja2
weasyprint
fpdf
matplotlib
pandas
numpy
tk

